create PositionSeeder

+++ database/migrations/2023_03_22_153324_create_positions_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->string('indirizzo');
             $table->unsignedTinyInteger('N_civico');
-            $table->float('Latitudine');
-            $table->float('Longitudine');
+            $table->double('Latitudine');
+            $table->double('Longitudine');
             $table->string('città', 50);
             $table->string('Nazione',20);
             $table->timestamps();
